+++ src/main/java/com/fadeevaaa/healthyeating/usermodule/service/UserServiceImpl.java
@@ -16,12 +16,12 @@ public class UserServiceImpl implements UserService {
     }
 
     public User create(User user) {
-        double dailyNorm = calculateDailyNorm(user.getWeight(), user.getHeight(), user.getAge());
+        short dailyNorm = calculateDailyNorm(user.getWeight(), user.getHeight(), user.getAge());
         user.setDailyNorm(dailyNorm);
         return userRepository.save(user);
     }
 
-    public double calculateDailyNorm(short weight, short height, byte age) {
-        return 88.36 + (13.4 * weight) + (4.8 * height) - (5.7 * age);
+    public short calculateDailyNorm(short weight, short height, byte age) {
+        return (short) (88.36 + (13.4 * weight) + (4.8 * height) - (5.7 * age));
     }
 }
